% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/URIparsing.R
\name{docdb_uri_base}
\alias{docdb_uri_base}
\alias{docdb_uri_base_parse}
\title{Construct a base URI for accessing DocumentDB}
\usage{
docdb_uri_base(database_account)

docdb_uri_base_parse(uri)
}
\arguments{
\item{database_account}{Name of database account}

\item{uri}{Unique Resource}
}
\value{
base_uri Constructed base URI

parsed Boolean, true where URI
}
\description{
Construct a base URI for accessing DocumentDB

Parse a base URI for accessing DocumentDB
}
\section{Functions}{
\itemize{
\item \code{docdb_uri_base}: 

\item \code{docdb_uri_base_parse}: 
}}
\examples{
docdb_uri_base("myDB")
docdb_uri_base_parse("https://myDB.documents.azure.com")
}

